current perf issues:
updateMatrixWorld - 20%
multiplyMatrices - 17%
renderObject - 14%
projectObject - 13%
updateSelf (leaf) - 10%
Object.traverse - 8%

this is mostly caused by very large numbers of Object3D instances each
with their own position/scale/rotation and deep nesting.

This is because, for leaves, every vein in the leaf is a Bone - this is up to
3000 bones per leaf.

We have 82,000 Objects!!! each one live updating the matrix world every
frame.